cleanup: move SendToPeer to shell/executor.go.
